delete_post and update_posts updated

+++ app/routers/post.py
@@ -52,6 +52,9 @@ def delete_post(id: int, db: Session = Depends(database.get_db), current_user: i
     if post == None:
         raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail= f"Post {id} was not found")
     
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code= status.HTTP_403_FORBIDDEN, detail= "Not authorized to perform current action")
+
     post_query.delete(synchronize_session = False)
     db.commit()
     return Response(status_code= status.HTTP_204_NO_CONTENT)
@@ -67,6 +70,10 @@ def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends
     post = post_query.first()
     if post == None:
         raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail= f"Post {id} was not found")
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code= status.HTTP_403_FORBIDDEN, detail= "Not authorized to perform current action")
+    
     post_query.update(updated_post.dict(),synchronize_session = False)
     db.commit()
     #return {"data": updated_post}
